+++ docs/en/changes/changes.md
@@ -111,6 +111,9 @@
 * Fix the malfunctioning alarm feature of MAL metrics due to unknown metadata in L2 aggregate worker.
 * Make MAL percentile align with OAL percentile calculation.
 * Update Grafana dashboards for OAP observability.
+* BanyanDB: fix query `getInstance` by instance ID.
+* Support the go agent(0.7.0 release) bundled pprof profiling feature. 
+
 
 #### UI